+++ packages/flutter_pptx/lib/src/assets.dart
@@ -3,16 +3,17 @@ import 'package:flutter/services.dart';
 
 class FlutterAssets extends Assets {
   @override
-  Future<Uint8List?> getImageData(String path) async {
-    if (hasCache(path)) return getCache(path);
+  Future<Uint8List?> getImageData(ImageReference img) async {
+    final path = img.path;
+    if (hasCache(img)) return getCache(img);
     try {
       final data = await rootBundle.load(path);
       final bytes = data.buffer.asUint8List();
-      setCache(path, bytes);
+      setCache(img, bytes);
       return bytes;
     } catch (e) {
       // debugPrint('Failed to load image: $path');
     }
-    return super.getImageData(path);
+    return super.getImageData(img);
   }
 }