+++ packages/flutter_pptx/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+- Fix for images being duplicated in the same slide
+- Updating `dart_pptx` dependency to `0.1.2`
+
 ## 0.1.1
 
 - Fixing hidden files not uploading to pub.dev